anavi-miracle-controller-sw.ino: Dynamically set LEDs
Save LED type and color order during the initial configuration.
After that at boot dynamically set LED type and color order with
a spaghetti code to pass arguments to FastLED.addLeds. For the
moment the configuration for color order is not really taken into
account. Only a limited number of configurations are supported:
WS2812B (default), WS2812, WS2811 and NeoPixels

Altermatively, during compile, if LED_TYPE1 is defined the type
and color orders of both LEDs will be set statically and the saved
configurations will be ignored.

Appropriate messages will be printed in the serial output
depending on the configuration.

Signed-off-by: Leon Anavi <leon@anavi.org>