Add the "mqtt" effect, where individual leds can be set via MQTT
For example, this message set LED 0 to a dim red and LED 1 to a bright
blue.  The other LEDs will retain their color.

    {effect:"mqtt",pixels:[{n:0,r:128,g:0,b:0},{n:1,r:0,g:0,b:255]}

Due to limitations in the MQTT library, you can only change a few LEDs
at a time.  Due to the speed limits of MQTT this method is not
appropriate to create dynamic patterns, but it can be used for status
displays or to try out a patterns before coding them in the firmware.